requirements.txt
beautifulsoup4>=4.9.3
playwright
pandas
requests>=2.26.0
streamlit>=1.10.0
altair>=4.2.0
pymysql
python-dotenv